Implement helper to clone a prelude rule with overrides
Summary:
There's a lot of macro logic built up for defining prelude rules,
especially those imported from buck1. It is useful to be able to reuse these
implementations but with slight tweaks, like, implementing `toolchain_command_alias`
which is a variant of `command_alias` but as a toolchain rule and changing the
type of the `exe` parameter to `exec_dep`. Actually doing this though is
almost impossible because of all the computation and logic that goes into
defining the original `command_alias`. So, expose an additional function
in rules.bzl called `clone_rule` which instead accepts the desired overrides
and invokes the exact same logic with the same inputs as the rule to be cloned,
just applying the extra bits on top.

Reviewed By: stepancheg

Differential Revision: D48768248

fbshipit-source-id: f1432af35088aafd7649a8b7ff36b3f18478dfcd