Refactor out the `shared_lib` creation function
Summary:
Unfortunately, getting the transitive dep caching moved to a
top-level, shared function is a bit non-trivial. We probably need to
add a provider to the targets to cache the values but it's maybe not
worth it.

This means we don't actually validate that the relinker is working over prebuilts with dependencies which is maybe unsafe.

Reviewed By: cjhopman

Differential Revision: D52086654

fbshipit-source-id: 81ab6ce7ba9d8005f2f24d8a46c3bfa1a7e385e0