Don't try to check cpu_filters when installing if we don't have native libraries
Summary:
We pass the `cpu_filters` to the installer so that the installer can check if the native libraries are compatible with the device that we are installing to.

If there are not native libraries, then that check doesn't make any sense, so don't pass the `cpu_filters` at all.

See discussion at https://fb.workplace.com/groups/695333970515221/posts/25561410293480911/?comment_id=25565325519756055&reply_comment_id=25568488129439794

Reviewed By: navidqar

Differential Revision: D52161496

fbshipit-source-id: b7abe25922cb1207a61164b2b3c2c12cd854e356