rules: Start cleaning up `LinkStyle` usage
Summary:
`LinkStyle` is deprecated and has largely been removed from the C++ rules in favor of `Linkage` (the attribute on libraries), `LinkStrategy` (the key for fields in `MergedLinkInfo` and the attribute on binaries) and `LibOutputStyle` (self-explanatory).

Rust should follow suit. This diff converts the usage of `dep_link_style`, specifically the value in the `BuildParams` and goes from there. The use of `LinkStrategy` for this value is correct, as `LinkStrategy` is primarily about the way dependencies are treated, and not the artifact being compiled itself

Reviewed By: chatura-atapattu

Differential Revision: D52450316

fbshipit-source-id: dcbf55a364eedcc0a2e97f33a0c03056814ed971