rules: There should be only one `LinkStrategy`
Summary:
Mostly explained in the comment in the diff

Whether this diff is an improvement should be seen in conjunction with the next diff. The argument in favor of this change is that it's written in a way that aligns better with the C++ code, which is good both for exposing incompatibilities and for helping people familiar with C++ understand the Rust behavior.

Reviewed By: dtolnay

Differential Revision: D52450304

fbshipit-source-id: 275cb526593d52b9680b31d855eff4a5d9be432f