rules: Fix pre-stripped artifacts
Summary:
It's not clear to me if D49068069 was supposed to work on its own or only in conjunction with native unbundled deps. In any case, I can't reproduce it doing anything for opt builds, which is unsurprising given that the code added there is `LibOutputStyle("shared")` specific.

This diff moves it to the right branch

Reviewed By: dtolnay

Differential Revision: D52499870

fbshipit-source-id: ed536f56760274a528bfc6738e04dc3b1dea264b