rules: Add a comment
Summary: I really wish this were better documented. Link group libs are explicitly defined targets that act as the link group shared libraries - this is an alternative to auto link groups in which those are automatically created. Rust does not (and should not) support acting as such a target, but may appear in a link graph in which such targets appear and so does have to forward the providers

Reviewed By: dtolnay

Differential Revision: D52499871

fbshipit-source-id: dda816b31c53d3f670cfe3dcfdf2abef3cdaa6b1