enable package regex matching for group_mapping
Summary:
We would like to be able to define link group map by regex matching on a target package.
Today we can either to regex on labels or match target pattern by exact target or directory prefix.

But for cases such as dsi/logger/configs there are a lot of dircetories (~100K) under it which we would like to be able to split into a finer granularity groups.

The trigger for this is user reloc error that is coming from the logger.so itself after it was split into a seperate general link group : [post](https://fb.workplace.com/groups/linker.support/posts/1092376945262480/?comment_id=1093963691770472&reply_comment_id=1094532725046902)

Reviewed By: chatura-atapattu, athmasagar

Differential Revision: D51926276

fbshipit-source-id: a5ad537f84696c9d98a573e0efefcb009a8cc0c6