Add `validation_deps` field
Summary:
## Context

In Buck1, `genrule()` target which appear in the `deps` field of targets would get executed. This was utilised to attach various "validation" genrules which would run some logic.

In Buck2, such genrule targets would not be executed unless their outputs are explicitly requested, which for validation rules does not happen.

## Genrule Deps

We added the ability to execute all genrule `deps` of `apple_*` rules. While this approach works in making sure validation rules run, it has several downsides:

- **Performance**: it ends up slowing down builds because *not all* genrules are used for validation. Thus we end up with overbuilding and creating additional waiting on dependents.
- **Long Term State**: replicating buck1 behaviour is undesirable as a long-term state because of performance and usability implications.
  - For example, clearly separating validation deps allows us to handle them differently in the future (e.g., skipping validation in certain uses case, providing explicit `[validation]` subtargets etc).
  - Prevents unification of `apple_*` rules with `cxx_*` rules due to dependence on the behaviour.

## Overview

The solution is to have a separate `validation_deps` field on `apple_binary()`, `apple_library()`, `apple_bundle()` and `apple_test()`. Any targets in the `validation_deps` fields are *guaranteed* to have been built as part of the outputs of the respective rules.

This is affectively an opt-in approach to transitioning the pre-existing validation targets.

Reviewed By: blackm00n

Differential Revision: D52656976

fbshipit-source-id: 80e3f4d3865dccea6b6096c23beeeef5b4180530