add apple_xcuitest rule
Summary:
Add basic support for XCUITest runner application bundles. These need to package a test bundle in the PlugIns folder, along with their required frameworks in Frameworks.

For now we are duplicating the frameworks across the test runner and the test application, removing them would require some refactoring of `assemble_bundle` to take in an exclude list or similar.

Reviewed By: blackm00n

Differential Revision: D52547900

fbshipit-source-id: 523986ff39c556e4c07ae991882bef7a18be56f6