Work around command line length overflow in transitive_dependency_symlinks
Summary:
Previously, transitive dependencies were passed to this script in the form:

```
transitive_dependency_symlinks.py \
    --out-dir path/to/out \
    --artifact path/to/cratename ../../libprovisional.rlib \
    --artifact ...
```

This doesn't work because there can be a lot of them, overflowing the platform's command line length limit.

This diff passes transitive dependencies through a JSON file:

```
transitive_dependency_symlinks.py \
    --out-dir path/to/out \
    --artifacts path/to/artifacts.json
```

where artifact.json is an array of pairs, each an rlib and a file containing the crate name for it.

```
[
    ["../../libprovisional.rlib", "path/to/cratename"],
    ...
]
```

Reviewed By: diliop

Differential Revision: D52759696

fbshipit-source-id: d7c81c8cedc1ed31cb76fc4a8cda3b1c175ff7d7