[buck2] Support haskell sources with irregular module name
GHC is expecting an interface file in the search path at a location determined
from the module name as given by the module-id from the `module` statement.

In the build, a module's name is only inferred from the path of the source file,
which might not correctly reflect the actual module name as seen by GHC.

Reading the dependency information, we can determine the correct module name for a
given source file and use that information to fix the mapping of module name to
`_Module` information.