Expose new make_swift_interface.py tool
Summary: In order to implement the `[swift-interface]` subtarget, we need to convert flags meant for `swiftc` into flags that `swift-ide-test` understands. We also need a driver for `swift-ide-test` which writes the `stdout` output to a file (as `swift-ide-test` has no `--out` flag). This diff introduces a new tool which accomplishes both tasks: it accepts the `swiftc` arguments meant to shuffle over to `swift-ide-test` & a path to `swift-ide-test` and invokes `swift-ide-test` writing `stdout` to the specified out file.

Reviewed By: drodriguez

Differential Revision: D53098983

fbshipit-source-id: 4512b1cd959ad5cec7659b9583e8a3dd104988bd