Sanitizers: add sanitizer fields to `cxx_toolchain()`
Summary:
Adds two fields to `cxx_toolchain()`, so that the correct sanitizer runtime libs can be referenced and bundled as required.

- `sanitizer_runtime_dir`:
- `sanitizer_runtime_enabled`: controls whether the sanitizer runtime directory would be referenced during linking using `rpath` linker flags and whether the runtime would be embedded in any standalone app bundles (e.g., macOS applications).

The directory and flag to enable are separated out, for easier debugging and visibility (vs having an empty `sanitizer_runtime_dir` control both).

`sanitizer_runtime_dir` is a dep and not an exec-dep because that allows the runtime to be customised on a per-target platform basis (e.g., macOS vs iOS get different runtime dylibs).

Reviewed By: blackm00n

Differential Revision: D53264639

fbshipit-source-id: ea987ff0188b3c9e7d7f3287bb928ed95e63f2d5