packagedb_args_thin -> packagedb_args
The separation is redundant. The field exposed_packages_artifacts
already carries all the interfaces, objects, and libraries; and these
are already inputs where packagedb_args was an input. So, there is no
need for a separte package_args_thin. Instead, package_args only carries
the package-db flag and inputs, and exposed_packages_artifacts carries
the interfaces, objects, and libraries.