profile selection for adhoc signing
Summary:
profile is not yet copied to bundle, just selection is running, so
functionality not yet finished. Selection implementation is a bit different
from buck1 — while buck1 was not completely ignoring certificate identities, in
buck2 we limit the list of available certificates to the one set directly in
target definition though we check this certificate compatibility when selecting
provisioning profile. That is more tight approach and more straight forward for
implementation.

Reviewed By: milend

Differential Revision: D53611185

fbshipit-source-id: 501f745d4356c154d61d8f8d08c9b5d5262aedc5