Suppress error for Apple framework pcm
Summary:
> buck-out/v2/gen/fbsource/0d0ff24630454274/xplat/toolchains/apple/__xcode_15.3.0_15e5178i-iphonesimulator-sdk_genrule__/iPhoneSimulator17.4.sdk/System/Library/Frameworks/AssetsLibrary.framework/Headers/ALAssetsLibrary.h:19:2: error: AssetsLibrary will be removed from the iOS SDK in the next major release [-Werror,-W#warnings]
#warning AssetsLibrary will be removed from the iOS SDK in the next major release

This is actually a warning from the compilation of the AssetsLibrary pcm in the apple toolchain cache generation. Pass -Wno-error=#warnings to those builds to suppress this error. We should still get it when importing these headers in our own code, but those are being systematically removed already T179144253 by the FoA Deprecated API workstream.

Reviewed By: rmaz

Differential Revision: D53742183

fbshipit-source-id: a603183a18826786b75af063c8918a5b930d9431