Only generate separate R.java for strings and ids if we're doing pre-dexing
Summary:
If we're pre-dexing, we want to write any strings and ids we have to a separate Java library, since when we're pre-dexing we have a limit of 65K fields and that gives us more headroom.

If we're not pre-dexing, then we don't need to do that since we're just going to pass every Java library to `d8` at the same time anyway.

This now matches `buck1` behavior, where this writing to a separate Java library is done in `SplitUberRDotJava` and is only invoked for pre-dexing.

Reviewed By: navidqar

Differential Revision: D54065428

fbshipit-source-id: b2e28e8ed53cecf5fdb75bac90dc9e44d669f981