Ensure that target json files are used to scrub SelectedDebugPaths.txt
Summary:
Context: https://fb.workplace.com/groups/buck2apple/posts/724583839460778/

TL:DR; Users are complaining that they are stopping in the debugger for every single breakpoint in the app.

The root cause is that the `focused_targets.json` file (which is populated by the iOS OnDemand tool) is not used to filter the `SelectedDebugPaths.txt` in `apple_bundle.bzl` it only looks at variable. As a result, almost all paths in the binary are included (yikes).

This diff updates the `apple_bundle` logic so that it can parse the `focused_targets.json` and use it to further scrub the `SelectedDebugPaths.txt`.

Reviewed By: milend

Differential Revision: D54032576

fbshipit-source-id: eb0d4b6ff0e33e633b72c8e9ce95d1bc146092f8