[buck2] Prevent Makefile clobbering
`ghc -M` will write to a file called `Makefile` in the working directory
when `-dep-makefile` is not set, even if `-dep-json` is set.

Since local Buck2 builds are not isolated, this means that the user's
repository directory will be clobbered.

This change passes `-dep-makefile` to direct GHC's output towards a
temporary file to avoid this.