Add support for validation_deps to android_binary
Summary:
Now that `validation_deps` is not coupled (by name) to Apple targets, we can reuse it in `android_binary` so that build-time validation rules (e.g. dependency rules) can be added to Android apps as well.

This could also apply to `android_library` etc. but I don't have a need for it there and I don't know if we'd really want to add them, as it would create a lot of additional targets in the graph. So I'm only adding it to `android_binary`.

Reviewed By: IanChilds

Differential Revision: D54097352

fbshipit-source-id: e8db6d34d36f75ec76914bf4d8830e5adedb9ac3