add native support for modules with non-asset cxx_library with gating
Summary:
This stack of diffs is config-erized version of D53624135

Buck was written to support asset native libs but not non-asset native libs. We've had a need for this with helium that we supported via hacks in our voltron postprocessor.

In preparation to enable Voltron for WA, we need to build a similar solution as they do not use asset libraries at all.
- We can force WA to use asset libs
- We can port over existing postprocessor hack to WA as well as write similar support for AABs
- We can add support to Buck to handle this correctly (we're going with this)
The expected behavior for non-asset native libs is for the .so artifact to be placed in <module>/lib/<cpu>/ in both the APK as well as the AAB. For AABs, we should also expect a native.pb file to be generated.

=====

when this is fully rolled out, consider reverting this stack and just pushing D53624135 for clean up

Reviewed By: IanChilds

Differential Revision: D53968378

fbshipit-source-id: b423fd60e707bb753ef575f1bd37e48705ffd47a