Add subtargets for individual merged native libraries
Summary:
For debugging It's useful to build individual merged (or unmerged) libraries as they are built for an app instead of waiting for all of them to build.

Due to the fact that the final merged outputs are produced as dynamic outputs (determined after analysis), we aren't able to match subtarget names to final outputs exactly.
- Split groups are not known at analysis time, so we expose a subtarget for each merge group, that builds all the split groups for a merge group
- The libraries that are unmerged in the final output are unknown, so building ones that end up in merge group will build that merge group instead
- Some merge groups aren't in the final output because they don't include anything, building these builds the native merge debug output

Instead of linking to alternative outputs, we could trigger an action that always fails when building subtargets that aren't in the final output, but the error that would be produced is a bit hard to parse and not ergonomic to generate, so this seems a bit better.

Reviewed By: IanChilds

Differential Revision: D53834064

fbshipit-source-id: c633698605833c8664a634ee03a29572ccbc53f7