Add native_library_merge_linker_args for android_binary
Summary:
`native_library_merge_linker_args` allows specifying a map that looks like `{"libmerged.so": ["--linker-arg"]}` so pass linker args to final merged outputs which can be merge group or split group names.

To supply mip profile data to the linker, we need a way to specify linker args for merged shared libraries. Currently mip profiles use some automation to identify a representative library in each shared library, and pass the args to individual libraries, but it's a bit fragile and depends on debug outputs.

Reviewed By: IanChilds

Differential Revision: D53821751

fbshipit-source-id: 7ea420b5c3f2c115953858435c3ac68b5baf844c