Give MergeAssetsExecutableMain the ability to write modular assets APKs
Summary:
This changes `MergeAssetsUtils` to accept a map of modules-to-assets.

Assets in the "root" module are treated the same as they are now (i.e. put into the APK that is being merged). For assets in a non-root module, we create a new `assets.ap_` and put it into `<module_assets_apks_dir>/module`.

Reviewed By: bchang7

Differential Revision: D54419324

fbshipit-source-id: 4da870f53a4842283697f0785cf6aba77cb44865