Migrate non-fbcode sysroot mechanism to new buck2 one
Summary:
Following the fbcode buckified sysroot migration, this diff moves the non-fbcode usages of rust in buck2 to our new toolchain-supported mechanism.

With this change, the rust_sysroot_decorator is unnecessary in buck2, and remains around just to support the existing buck1 usages on android and iOS. Once those migrate in Q3, we can delete the decorator entirely.

Reviewed By: zertosh, JakobDegen

Differential Revision: D54130104

fbshipit-source-id: 0ee703008fe9fbdd8b706215eccbe526faf1f4c0