Put module assets into module/assets for Android Bundles
Summary:
Now, if we have an Android Bundle and a Voltron mapping, we pass a map of module-to-assets so that `merge_assets` will create a separate `assets.ap_` in our "module assets apks dir".

If we produce that dir, then we pass it to `AndroidBundleBuilderExecutableMain` and for each module in the bundle, we look for an `assets.ap_` in the dir. If it exists, we use that when creating the "module" rather than just using a "fake" resources APK.

Reviewed By: bchang7

Differential Revision: D54417876

fbshipit-source-id: 430ddad628a9f56d93864eec218600fe79e0dc8d