Fail if exclude-libs,ALL is used
Summary:
This flag frequently causes buck2 to produce builds that crash at runtime. Let's make it a build time failure to clean up remaining uses, and prevent it from being re-introduced.

This could be done in the macro layer but it should be cheap enough to check it for any android_binary. Since this affects shared library dependencies of binary builds only the flag can still be used by AARs or standalone library builds where it's possible this may still have some use (or may just be harder to clean up).

Reviewed By: nlutsenko

Differential Revision: D54439931

fbshipit-source-id: 67ef6e28a3e1a9b662659b977342dc6cc76ab6d6